Make the workspace model conform ModelConvertible
